src/backend/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.2
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.0.2
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.1
werkzeug==3.1.3
zipp==3.21.0
